requirements.txt
rembg==2.0.50
Pillow==9.5.0
numpy==1.24.3
onnxruntime==1.15.1
